move Pattern to gamut
