+++ packages/styleguide/stories/Core/Atoms/Pattern.stories.mdx
@@ -1,4 +1,4 @@
-import { Pattern } from '@codecademy/gamut-labs/src';
+import { Pattern } from '@codecademy/gamut/src';
 import { Meta, Canvas, Props, Story } from '@storybook/addon-docs/dist/blocks';
 
 <Meta
